+++ examples/paytechuz_django/payment/views.py
@@ -1,5 +1,9 @@
 # views.py
-from paytechuz.integrations.django.views import BasePaymeWebhookView, BaseClickWebhookView
+from paytechuz.integrations.django.views import (
+    BasePaymeWebhookView,
+    BaseClickWebhookView,
+    BaseAtmosWebhookView
+)
 
 from shop.models import Order
 
@@ -17,6 +21,18 @@ class PaymeWebhookView(BasePaymeWebhookView):
 
 
 class ClickWebhookView(BaseClickWebhookView):
+    def successfully_payment(self, params, transaction):
+        order = Order.objects.get(id=transaction.account_id)
+        order.status = 'paid'
+        order.save()
+
+    def cancelled_payment(self, params, transaction):
+        order = Order.objects.get(id=transaction.account_id)
+        order.status = 'cancelled'
+        order.save()
+
+
+class AtmosWebhookView(BaseAtmosWebhookView):
     def successfully_payment(self, params, transaction):
         order = Order.objects.get(id=transaction.account_id)
         order.status = 'paid'