+++ examples/paytechuz_django/shop/models.py
@@ -14,6 +14,7 @@ class Order(models.Model):
     PAYMENT_TYPE_CHOICES = (
         ('payme', 'Payme'),
         ('click', 'Click'),
+        ('atmos', 'Atmos'),
     )
 
     product_name = models.CharField(max_length=255)