+++ src/integrations/django/__init__.py
@@ -1,4 +1,0 @@
-"""
-Django integration for PayTechUZ.
-"""
-default_app_config = 'paytechuz.integrations.django.apps.PaytechuzConfig'