+++ src/integrations/django/migrations/__init__.py
@@ -1,3 +1,0 @@
-"""
-Django migrations for PayTechUZ.
-"""